backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
tensorflow==2.13.0
numpy==1.23.5
Pillow==8.4.0
python-dotenv==0.19.0


